Refactor EventPage types
