+++ apps/eventer_web/assets/js/src/features/eventPage/useChannel.ts
@@ -1,19 +1,19 @@
-import { Option, Some, None } from 'funfix';
+import { None, Option, Some } from 'funfix';
+import { Channel, Socket } from 'phoenix';
 import { useEffect, useState } from 'react';
-import { Socket, Channel } from 'phoenix';
-import {
-  mapResponseEventToStateEvent,
-  addUserToParticipants,
-  moveToExParticipants,
-  updateStateEvent,
-  updateStateDecision,
-  resolveStateDecision,
-  openStateDiscussion,
-  stateEventT,
-  addStateDecision,
-  removeStateDecision,
-} from './stateTransformations';
 import { useParams } from 'react-router-dom';
+import {
+  addStateDecision,
+  addUserToParticipants,
+  mapResponseEventToStateEvent,
+  moveToExParticipants,
+  openStateDiscussion,
+  removeStateDecision,
+  resolveStateDecision,
+  updateStateDecision,
+  updateStateEvent,
+} from './stateTransformations';
+import { stateEventT } from './types';
 
 type useChannelT = (
   token: string,
